+++ apps/api/src/participation/participation.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateParticipationDto } from './dto/create-participation.dto';
 import { UpdateParticipationDto } from './dto/update-participation.dto';
 import { PrismaService } from 'src/prisma.service';
@@ -15,10 +15,26 @@ export class ParticipationService {
   }
 
   findOne(id: string) {
-    return this.prisma.participation.findUnique({ where: { id } });
+    const participation = this.prisma.participation.findUnique({
+      where: { id },
+    });
+
+    if (!participation) {
+      throw new NotFoundException(`Participation with id ${id} not found`);
+    }
+
+    return participation;
   }
 
   update(id: string, updateParticipationDto: UpdateParticipationDto) {
+    const participation = this.prisma.participation.findUnique({
+      where: { id },
+    });
+
+    if (!participation) {
+      throw new NotFoundException(`Participation with id ${id} not found`);
+    }
+
     return this.prisma.participation.update({
       where: { id },
       data: updateParticipationDto,
@@ -26,6 +42,14 @@ export class ParticipationService {
   }
 
   remove(id: string) {
+    const participation = this.prisma.participation.findUnique({
+      where: { id },
+    });
+
+    if (!participation) {
+      throw new NotFoundException(`Participation with id ${id} not found`);
+    }
+
     return this.prisma.participation.delete({ where: { id } });
   }
 }
