+++ apps/api/src/sport/sport.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateSportDto } from './dto/create-sport.dto';
 import { UpdateSportDto } from './dto/update-sport.dto';
 import { PrismaService } from 'src/prisma.service';
@@ -15,14 +15,32 @@ export class SportService {
   }
 
   findOne(id: string) {
-    return this.prisma.sport.findUnique({ where: { id } });
+    const sport = this.prisma.sport.findUnique({ where: { id } });
+
+    if (!sport) {
+      throw new NotFoundException(`Sport with id ${id} not found`);
+    }
+
+    return sport;
   }
 
   update(id: string, updateSportDto: UpdateSportDto) {
+    const sport = this.prisma.sport.findUnique({ where: { id } });
+
+    if (!sport) {
+      throw new NotFoundException(`Sport with id ${id} not found`);
+    }
+
     return this.prisma.sport.update({ where: { id }, data: updateSportDto });
   }
 
   remove(id: string) {
+    const sport = this.prisma.sport.findUnique({ where: { id } });
+
+    if (!sport) {
+      throw new NotFoundException(`Sport with id ${id} not found`);
+    }
+
     return this.prisma.sport.delete({ where: { id } });
   }
 }
