+++ apps/api/src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma.service';
@@ -17,12 +17,26 @@ export class UserService {
   }
 
   findOne(id: string) {
-    return this.prisma.user.findUnique({
+    const user = this.prisma.user.findUnique({
       where: { id },
     });
+
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
+    return user;
   }
 
   update(id: string, updateUserDto: UpdateUserDto) {
+    const user = this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
     return this.prisma.user.update({
       where: { id },
       data: updateUserDto,
@@ -30,6 +44,14 @@ export class UserService {
   }
 
   remove(id: string) {
+    const user = this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
     return this.prisma.user.delete({
       where: { id },
     });
